requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.9
filelock==3.16.1
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector==2.2.9
platformdirs==4.3.6
python-dateutil==2.9.0.post0
python-whois==0.9.5
requests==2.32.3
requests-file==2.1.0
six==1.17.0
tldextract==5.1.3
urllib3==2.3.0
virtualenv==20.28.0
Werkzeug==3.1.3
whois==1.20240129.2